Fix Java applications under Sway, add window rules
Java windows would previously display as a blank window, because the environment
variable _JAVA_AWT_WM_NONREPARENTING was not properly set, this fixes that.

Add window rules to float Matlab figures and inhibit idle when a Teams meeting
is visible.

Fix an issue where swayidle would not properly start on startup.